Move tests for edit and delete buttons to a new test file for RulesTable
